// src/clients/medusa/models/DatasetDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { EventDTO } from "./EventDTO";
import {
  EventDTOFromJSON,
  EventDTOFromJSONTyped,
  EventDTOToJSON,
  EventDTOToJSONTyped,
} from "./EventDTO";
import type { TableDTO } from "./TableDTO";
import {
  TableDTOFromJSON,
  TableDTOFromJSONTyped,
  TableDTOToJSON,
  TableDTOToJSONTyped,
} from "./TableDTO";

/**
 *
 * @export
 * @interface DatasetDTO
 */
export interface DatasetDTO {
  /**
   *
   * @type {Array<EventDTO>}
   * @memberof DatasetDTO
   */
  events: Array<EventDTO>;
  /**
   *
   * @type {Array<TableDTO>}
   * @memberof DatasetDTO
   */
  tables?: Array<TableDTO>;
}

/**
 * Check if a given object implements the DatasetDTO interface.
 */
export function instanceOfDatasetDTO(value: object): value is DatasetDTO {
  if (!("events" in value) || value["events"] === undefined) return false;
  return true;
}

export function DatasetDTOFromJSON(json: any): DatasetDTO {
  return DatasetDTOFromJSONTyped(json, false);
}

export function DatasetDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DatasetDTO {
  if (json == null) {
    return json;
  }
  return {
    events: (json["events"] as Array<any>).map(EventDTOFromJSON),
    tables:
      json["tables"] == null ? undefined : (json["tables"] as Array<any>).map(TableDTOFromJSON),
  };
}

export function DatasetDTOToJSON(json: any): DatasetDTO {
  return DatasetDTOToJSONTyped(json, false);
}

export function DatasetDTOToJSONTyped(
  value?: DatasetDTO | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    events: (value["events"] as Array<any>).map(EventDTOToJSON),
    tables:
      value["tables"] == null ? undefined : (value["tables"] as Array<any>).map(TableDTOToJSON),
  };
}