// src/clients/medusa/models/FieldDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { GenerationDTO } from "./GenerationDTO";
import {
  GenerationDTOFromJSON,
  GenerationDTOFromJSONTyped,
  GenerationDTOToJSON,
  GenerationDTOToJSONTyped,
} from "./GenerationDTO";

/**
 *
 * @export
 * @interface FieldDTO
 */
export interface FieldDTO {
  /**
   *
   * @type {string}
   * @memberof FieldDTO
   */
  name: string;
  /**
   *
   * @type {GenerationDTO}
   * @memberof FieldDTO
   */
  generation: GenerationDTO;
  /**
   *
   * @type {boolean}
   * @memberof FieldDTO
   */
  ignored?: boolean;
}

/**
 * Check if a given object implements the FieldDTO interface.
 */
export function instanceOfFieldDTO(value: object): value is FieldDTO {
  if (!("name" in value) || value["name"] === undefined) return false;
  if (!("generation" in value) || value["generation"] === undefined) return false;
  return true;
}

export function FieldDTOFromJSON(json: any): FieldDTO {
  return FieldDTOFromJSONTyped(json, false);
}

export function FieldDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): FieldDTO {
  if (json == null) {
    return json;
  }
  return {
    name: json["name"],
    generation: GenerationDTOFromJSON(json["generation"]),
    ignored: json["ignored"] == null ? undefined : json["ignored"],
  };
}

export function FieldDTOToJSON(json: any): FieldDTO {
  return FieldDTOToJSONTyped(json, false);
}

export function FieldDTOToJSONTyped(
  value?: FieldDTO | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    name: value["name"],
    generation: GenerationDTOToJSON(value["generation"]),
    ignored: value["ignored"],
  };
}